Show missing webview dependency view  (#95)
This change creates and presents a view that indicates WebView dependency may be missing on a user's system and they need to install it.
For rendering WebViews in Eclipse, WebView2 Runtime must be present on Windows devices and WebKit must be present on Linux devices.
The Eclipse SDK API does not provide a way to definitively detect the absence of these two dependencies on a user's system. With that limitation in mind, a workaround was found based on the official documentation here which states
`Note: If WebView2 backend is requested but the runtime isn't found, the Browser will automatically fall back to the Internet Explorer backend.`
The browser type is determined after it is created and a check has been added to validate it is of the expected type. If it's not of the expected type, we currently assume the WebView2 runtime dependency for Windows and WebKit for Linux is missing and will show this view to the user.
This check is performed as soon as the Login View or the Amazon Q Chat view is attempting to load up and will switch to the missing dependency view if conditions are met. The subsequent components that utilize Browser are not created for these views.