Add support for patching env variable for Path on Mac (#461)
When eclipse is launched through finder on MacOs, it is a known issue that it does not inherit env variables. This leads to MCP servers failing initialization since they rely on env variables like path for resolving relative paths.
This change launches a shell to get system env variables for Path and inject it into the node process in situations like this.