Support updating remote when ABAP files are edited (#495)
Ensures when Q edits an existing file associated with an ABAP ADT project, the update is communicated to the remote server. This is done by marking the editor for the file as dirty and triggering a save which in turn informs remote. When a file is not opened, it is opened in an invisible editor to do the same. For new files not associated with the current workspace, when Q creates a new one user will be informed to configure it appropriately as it cannot be configured for them by Q directly due to the limitations around ABAP ADT projects.